chore: tentative fix
Remove the changes introduced in the previous commit and update the
`MessageEditor.confirm_mention_for_selection` method in order to request
autoscroll from the editor twice, one of them delayed, as this appears
to fix the scrolling issue, regardless of the number of cursors used.

I'm not yet entirely sure why this is the case, so further investigation
is probably warranted in order to actually understand if this is the
best fix we can do right now, and why.